requirements.txt
# Python dependencies for the PinoySeoul Monitoring System

# Used for interacting with the Docker daemon
docker==7.1.0

# Used for making HTTP requests, primarily for the Google Chat webhook
requests==2.32.3

# Used for parsing the YAML configuration file
PyYAML==6.0.1

# Used for loading environment variables from .env files
python-dotenv==1.0.1
